Rename registry_ to registry; servlets/request_ to servlets/request
