+++ lib/hub.js
@@ -1,5 +1,5 @@
 var log = require('./log'),
-    Registry = require('./registry_'),
+    Registry = require('./registry'),
     server = require('./server'),
     config = require('./config');
