+++ lib/registry/Node.js
