+++ lib/registry/NodeSet.js
