+++ lib/registry/NodeSlot.js
