+++ lib/registry/Registry.js
