+++ lib/registry/Session.js
