+++ lib/registry/SessionQueue.js
