+++ lib/registry/SessionSet.js
