+++ lib/registry/index.js
