+++ lib/server.js
@@ -14,7 +14,7 @@ var notImplementedServlet = require('./servlets/notImplemented'),
     apiTestSessionServlet = require('./servlets/apiTestSession'),
     registerServlet = require('./servlets/register'),
     unregisterServlet = require('./servlets/unregister'),
-    requestServlet = require('./servlets/request_');
+    requestServlet = require('./servlets/request');
 
 var servletRoutes = {
     '/grid/api/hub': apiHubServlet,
